Move theme color extractor to Theme directory
